requirements.txt
python-gitlab>=2.10.0